feat: Implement password hashing
This commit integrates Flask-Bcrypt for password hashing in the seed.py file.

The passwords for Manager, Employee, and HR_Personel instances are now being hashed before they are stored in the database.

This enhances the security of the application by ensuring that actual passwords are not stored in the database.