feat: Add DeleteEmployee component
This component includes a form that allows an HR member to delete an employee from the database.

When the form is submitted, it makes a DELETE request to the Flask backend.

If the request is successful, it displays a success message. If there's an error, it displays an error message.