fix: Add Authorization Header to Axios Requests
This commit includes the addition of an Authorization header to the Axios PATCH requests in the handleApprove and handleDecline functions in the PendingLeaves component.
This was done to resolve a 401 (UNAUTHORIZED) error that was occurring due to the server expecting an authorization token in the request headers, which was not being provided.
The access token is now retrieved from local storage using the retrieve function and included in the headers of the requests.